chore: added signup form logic and moved interfaces to a single file
