--- src/modules/auth/types/SingUpType.ts
@@ -1,5 +1,0 @@
-export interface singUpType {
-  username: string
-  password: string
-  passwordconfirm: string
-}
